+++ infra/src/storage_account.tf
@@ -25,6 +25,16 @@ resource "azurerm_storage_queue" "request-review-poison" {
   storage_account_name = module.storage_account.name
 }
 
+resource "azurerm_storage_queue" "request-publication" {
+  name                 = "request-publication"
+  storage_account_name = module.storage_account.name
+}
+
+resource "azurerm_storage_queue" "request-publication-poison" {
+  name                 = "request-publication-poison"
+  storage_account_name = module.storage_account.name
+}
+
 resource "azurerm_storage_queue" "request-history" {
   name                 = "request-history"
   storage_account_name = module.storage_account.name